requirements.txt
flask
flask-socketio
flask-cors
pillow
secrets
pydrive